Allow the user to run predictions on any of their fitbit metrics

+++ Model/Point.cs
@@ -3,6 +3,6 @@
     public class Point
     {
         public long x { get; set; }
-        public long y { get; set; }
+        public double y { get; set; }
     }
 }
